Fix 'leave game' modal, add launch screen
No longer presents modal at the end of a round, also restarts the
game after exiting and coming back to a new round. Our launch
screen file also got deleted, so I added it back.

Checks if the end of the round screen is presented before showing
the modal, and restarts the countdown text before the countdown
appears.